rearchitected main loop.  The data is no longer combined inside of this application.  Instead we save every data source to its own table in the Tracking Database and update the locations using SQL.
The updates are done via stored procedure.

I also modified the process that updates the unit location history to stop saving so many duplicate locations.  Now the system compares the lat/long at a precision of 4 decimal places (~11m resolution).  If they are different, it will add it to the history.  Otherwise it will update the location once every 5 minutes.

Last, I found that the process that runs this application was scheduled to start at 5:20 AM rather than 5 AM.  So I made some changes so that it would stop at 5:18 AM.